+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 ## Unreleased
 
 
+## 1.12.0
+### Added
+- slug model / serializer mixins
+- extendible health check
+- no_button_center modal type
+- continue_center modal type
+### Fixed
+- auth redirect in standalone auth mode
+- fixes issue with meta data cache expiry if expiry was set to never and api throttling was encountered
+
+
 ## 1.11.0
 ### Added
 - adds django-cors-headers to the project and the django default settings